+++ src/Oro/Bundle/SecurityBundle/AccessRule/Expr/Value.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace Oro\Bundle\SecurityBundle\AccessRule\Expr;
+
+use Oro\Bundle\SecurityBundle\AccessRule\Visitor;
+
+/**
+ * The access rule expression for a static value.
+ */
+class Value implements ExpressionInterface
+{
+    /** @var mixed */
+    private $value;
+
+    /**
+     * @param mixed $value
+     */
+    public function __construct($value)
+    {
+        $this->value = $value;
+    }
+
+    /**
+     * Returns the value.
+     *
+     * @return mixed
+     */
+    public function getValue()
+    {
+        return $this->value;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function visit(Visitor $visitor)
+    {
+        return $visitor->walkValue($this);
+    }
+}